python/src/requirements.txt
pandas
flask_restx==1.0.5
flask==2.3.3
certifi
mysql-connector-python
python-dateutil